refactor mongodb importer
