+++ local-drivers/mongodb/mongodb-sync-v4-driver/src/main/java/io/flamingock/oss/driver/mongodb/sync/v4/internal/MongoSync4Engine.java
@@ -33,7 +33,7 @@ import io.flamingock.core.local.LocalExecutionPlanner;
 import io.flamingock.core.local.TransactionManager;
 import io.flamingock.core.transaction.TransactionWrapper;
 import io.flamingock.core.engine.audit.importer.ImporterModule;
-import io.flamingock.importer.mongodb.sync4.MongoImporterReader;
+import io.flamingock.importer.mongodb.MongoImporterReader;
 import org.bson.Document;
 
 import java.util.Optional;
